Move model classes out of dto.model package to just dto
