+++ dto/src/main/java/org/jboss/pnc/dto/Build.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.dto.model;
+package org.jboss.pnc.dto;
 
 import org.jboss.pnc.enums.BuildCoordinationStatus;
