+++ dto/src/main/java/org/jboss/pnc/dto/BuildEnvironment.java
@@ -15,42 +15,49 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.dto.model;
+package org.jboss.pnc.dto;
 
-import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
-import lombok.Builder;
-import lombok.Data;
-
-import org.jboss.pnc.constants.Patterns;
 import org.jboss.pnc.dto.validation.groups.WhenCreatingNew;
 import org.jboss.pnc.dto.validation.groups.WhenUpdating;
+import org.jboss.pnc.enums.SystemImageType;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Null;
-import javax.validation.constraints.Pattern;
+
+import java.util.Map;
+
+import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
+
+import lombok.Builder;
+import lombok.Data;
 
 /**
  *
- * @author Jakub Bartecek &lt;jbartece@redhat.com&gt;
+ * @author Honza Brázdil &lt;jbrazdil@redhat.com&gt;
  */
 @Data
-@Builder(builderClassName = "Builder", builderMethodName = "refBuilder")
-public class ProductRef implements DTOEntity {
+@Builder(builderClassName = "Builder")
+public class BuildEnvironment implements DTOEntity {
+
     @NotNull(groups = WhenUpdating.class)
     @Null(groups = WhenCreatingNew.class)
-    protected final Integer id;
+    private final Integer id;
 
-    protected final String name;
+    private final String name;
 
-    protected final String description;
+    private final String description;
 
-    @NotNull(groups =  { WhenCreatingNew.class, WhenUpdating.class })
-    @Pattern(regexp = Patterns.PRODUCT_ABBREVIATION, groups = { WhenCreatingNew.class, WhenUpdating.class })
-    protected final String abbreviation;
+    private final String systemImageRepositoryUrl;
 
-    protected final String productCode;
+    @NotNull(groups = {WhenCreatingNew.class, WhenUpdating.class})
+    private final String systemImageId;
 
-    protected final String pgmSystemName;
+    private final Map<String, String> attributes;
+
+    @NotNull(groups = {WhenCreatingNew.class, WhenUpdating.class})
+    private final SystemImageType systemImageType;
+
+    private final boolean deprecated;
 
     @JsonPOJOBuilder(withPrefix = "")
     public static final class Builder {
