+++ dto/src/main/java/org/jboss/pnc/dto/GroupBuildRef.java
@@ -15,11 +15,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.dto.model;
+package org.jboss.pnc.dto;
 
 import org.jboss.pnc.dto.validation.groups.WhenCreatingNew;
 import org.jboss.pnc.dto.validation.groups.WhenUpdating;
-import org.jboss.pnc.enums.BuildType;
+import org.jboss.pnc.enums.BuildStatus;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Null;
@@ -37,27 +37,19 @@ import lombok.Data;
  */
 @Data
 @Builder(builderClassName = "Builder", builderMethodName = "refBuilder")
-public class BuildConfigurationRevisionRef implements DTOEntity {
+public class GroupBuildRef implements DTOEntity {
 
     @NotNull(groups = WhenUpdating.class)
     @Null(groups = WhenCreatingNew.class)
     protected final Integer id;
 
-    protected final Integer rev;
+    protected final Instant startTime;
 
-    protected final String name;
+    protected final Instant endTime;
 
-    protected final String description;
+    protected final BuildStatus status;
 
-    protected final String buildScript;
-
-    protected final String scmRevision;
-
-    protected final Instant creationTime;
-
-    protected final Instant modificationTime;
-
-    protected final BuildType buildType;
+    protected final Boolean temporaryBuild;
 
     @JsonPOJOBuilder(withPrefix = "")
     public static final class Builder {
