+++ dto/src/main/java/org/jboss/pnc/dto/ProductReleaseRef.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.dto.model;
+package org.jboss.pnc.dto;
 
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 import lombok.Builder;
@@ -23,7 +23,7 @@ import lombok.Data;
 
 import org.jboss.pnc.dto.validation.groups.WhenCreatingNew;
 import org.jboss.pnc.dto.validation.groups.WhenUpdating;
-import org.jboss.pnc.enums.MilestoneReleaseStatus;
+import org.jboss.pnc.enums.SupportLevel;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Null;
@@ -35,20 +35,21 @@ import java.time.Instant;
  * @author Jakub Bartecek &lt;jbartece@redhat.com&gt;
  */
 @Data
-@Builder(builderClassName = "Builder")
-public class ProductMilestoneRelease implements DTOEntity {
-
+@Builder(builderClassName = "Builder", builderMethodName = "refBuilder")
+public class ProductReleaseRef implements DTOEntity {
     @NotNull(groups = WhenUpdating.class)
     @Null(groups = WhenCreatingNew.class)
     protected final Integer id;
 
-    protected final MilestoneReleaseStatus status;
+    protected final String version;
 
-    protected final String log;
+    protected final SupportLevel supportLevel;
 
-    protected final Instant endDate;
+    protected final Instant releaseDate;
 
-    protected final Instant startingDate;
+    protected final String downloadUrl;
+
+    protected final String issueTrackerUrl;
 
     @JsonPOJOBuilder(withPrefix = "")
     public static final class Builder {
