+++ dto/src/main/java/org/jboss/pnc/dto/ProductVersionRef.java
@@ -15,30 +15,39 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.dto.model;
+package org.jboss.pnc.dto;
 
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 import lombok.Builder;
 import lombok.Data;
+
+import org.jboss.pnc.constants.Patterns;
 import org.jboss.pnc.dto.validation.groups.WhenCreatingNew;
 import org.jboss.pnc.dto.validation.groups.WhenUpdating;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Null;
+import javax.validation.constraints.Pattern;
+
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  *
  * @author Jakub Bartecek &lt;jbartece@redhat.com&gt;
  */
 @Data
-@Builder(builderClassName = "Builder")
-public class User implements DTOEntity {
-
+@Builder(builderClassName = "Builder", builderMethodName = "refBuilder")
+public class ProductVersionRef implements DTOEntity {
     @NotNull(groups = WhenUpdating.class)
     @Null(groups = WhenCreatingNew.class)
     protected final Integer id;
 
-    protected final String username;
+    @NotNull(groups = {WhenCreatingNew.class, WhenUpdating.class})
+    @Pattern(message="The version should consist of two numeric parts separated by a dot" , regexp=Patterns.PRODUCT_STREAM_VERSION, groups = {WhenCreatingNew.class, WhenUpdating.class})
+    protected final String version;
+
+    protected final Map<String, String> attributes = new HashMap<>();
 
     @JsonPOJOBuilder(withPrefix = "")
     public static final class Builder {
