+++ dto/src/main/java/org/jboss/pnc/dto/Project.java
@@ -15,14 +15,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.dto.model;
+package org.jboss.pnc.dto;
 
-import org.jboss.pnc.enums.BuildType;
+import org.jboss.pnc.dto.validation.constraints.RefHasId;
+import org.jboss.pnc.dto.validation.groups.WhenCreatingNew;
+import org.jboss.pnc.dto.validation.groups.WhenUpdating;
 
-import java.time.Instant;
-import java.util.Map;
+import java.util.List;
 
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
+
 import lombok.Data;
 
 /**
@@ -30,23 +32,18 @@ import lombok.Data;
  * @author Honza Brázdil &lt;jbrazdil@redhat.com&gt;
  */
 @Data
-public class BuildConfigurationRevision extends BuildConfigurationRevisionRef {
+public class Project extends ProjectRef {
 
+    @RefHasId(groups = {WhenCreatingNew.class, WhenUpdating.class})
     private final RepositoryConfiguration repositoryConfiguration;
 
-    private final ProjectRef project;
-
-    private final BuildEnvironment environment;
-
-    private final Map<String, String> genericParameters ;
+    private final List<BuildConfigurationRef> buildConfigurations;
 
     @lombok.Builder(builderClassName = "Builder")
-    public BuildConfigurationRevision(RepositoryConfiguration repositoryConfiguration, ProjectRef project, BuildEnvironment environment, Map<String, String> genericParameters, Integer id, Integer rev, String name, String description, String buildScript, String scmRevision, Instant creationTime, Instant lastModificationTime, BuildType buildType) {
-        super(id, rev, name, description, buildScript, scmRevision, creationTime, lastModificationTime, buildType);
+    public Project(RepositoryConfiguration repositoryConfiguration, List<BuildConfigurationRef> buildConfigurations, Integer id, String name, String description, String issueTrackerUrl, String projectUrl) {
+        super(id, name, description, issueTrackerUrl, projectUrl);
         this.repositoryConfiguration = repositoryConfiguration;
-        this.project = project;
-        this.environment = environment;
-        this.genericParameters = genericParameters;
+        this.buildConfigurations = buildConfigurations;
     }
 
     @JsonPOJOBuilder(withPrefix = "")
