+++ dto/src/main/java/org/jboss/pnc/dto/TargetRepositoryRef.java
@@ -15,45 +15,41 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.dto.model;
+package org.jboss.pnc.dto;
+
+import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
+import lombok.Builder;
+import lombok.Data;
 
 import org.jboss.pnc.dto.validation.groups.WhenCreatingNew;
 import org.jboss.pnc.dto.validation.groups.WhenUpdating;
-import org.jboss.pnc.enums.BuildCoordinationStatus;
+import org.jboss.pnc.enums.RepositoryType;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Null;
 
-import java.time.Instant;
-
-import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
-
-import lombok.Builder;
-import lombok.Data;
-
 /**
  *
- * @author Honza Brázdil &lt;jbrazdil@redhat.com&gt;
+ * @author Jakub Bartecek &lt;jbartece@redhat.com&gt;
  */
 @Data
 @Builder(builderClassName = "Builder", builderMethodName = "refBuilder")
-public class BuildRef implements DTOEntity {
-
+public class TargetRepositoryRef implements DTOEntity {
     @NotNull(groups = WhenUpdating.class)
     @Null(groups = WhenCreatingNew.class)
     protected final Integer id;
 
-    protected final Instant submitTime;
+    @NotNull(groups = {WhenUpdating.class, WhenCreatingNew.class})
+    protected final Boolean temporaryRepo;
 
-    protected final Instant startTime;
+    @NotNull(groups = {WhenUpdating.class, WhenCreatingNew.class})
+    protected final String identifier;
 
-    protected final Instant endTime;
+    @NotNull(groups = {WhenUpdating.class, WhenCreatingNew.class})
+    protected final RepositoryType repositoryType;
 
-    protected final BuildCoordinationStatus status;
-
-    protected final String buildContentId;
-
-    protected final Boolean temporaryBuild;
+    @NotNull(groups = {WhenUpdating.class, WhenCreatingNew.class})
+    protected final String repositoryPath;
 
     @JsonPOJOBuilder(withPrefix = "")
     public static final class Builder {
