--- dto/src/main/java/org/jboss/pnc/dto/model/RepositoryConfiguration.java
@@ -1,61 +1,0 @@
-/**
- * JBoss, Home of Professional Open Source.
- * Copyright 2014-2018 Red Hat, Inc., and individual contributors
- * as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jboss.pnc.dto.model;
-
-import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
-import lombok.Builder;
-import lombok.Data;
-
-import org.hibernate.validator.constraints.NotBlank;
-import org.jboss.pnc.dto.validation.groups.WhenCreatingNew;
-import org.jboss.pnc.dto.validation.groups.WhenUpdating;
-
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Null;
-
-import org.jboss.pnc.dto.validation.constraints.SCMUrl;
-
-/**
- *
- * @author Jakub Bartecek &lt;jbartece@redhat.com&gt;
- */
-@Data
-@Builder(builderClassName = "Builder")
-public class RepositoryConfiguration implements DTOEntity {
-
-    @NotNull(groups = WhenUpdating.class)
-    @Null(groups = WhenCreatingNew.class)
-    protected final Integer id;
-
-    @NotBlank(groups = {WhenUpdating.class, WhenCreatingNew.class})
-    @SCMUrl(groups = {WhenUpdating.class, WhenCreatingNew.class})
-    protected final String internalUrl;
-
-    protected final String internalUrlNormalized;
-
-    @SCMUrl(groups = {WhenUpdating.class, WhenCreatingNew.class})
-    protected final String externalUrl;
-
-    protected final String externalUrlNormalized;
-
-    protected final Boolean preBuildSyncEnabled;
-
-    @JsonPOJOBuilder(withPrefix = "")
-    public static final class Builder {
-    }
-}
